Visual studio compatibility
This removes some inlines when using visual studio to compile, uisng TPT_NO_INLINE. It also fixes many other problems visual studio has with the code

+++ src/game/GameView.cpp
@@ -2069,7 +2069,7 @@ void GameView::OnDraw()
 ui::Point GameView::lineSnapCoords(ui::Point point1, ui::Point point2)
 {
 	ui::Point newPoint(0, 0);
-	float snapAngle = floor(atan2(point2.Y-point1.Y, point2.X-point1.X)/(M_PI*0.25)+0.5)*M_PI*0.25;
+	float snapAngle = floor(atan2((float)point2.Y-point1.Y, point2.X-point1.X)/(M_PI*0.25)+0.5)*M_PI*0.25;
 	float lineMag = sqrtf(pow((float)(point2.X-point1.X),2)+pow((float)(point2.Y-point1.Y),2));
 	newPoint.X = (int)(lineMag*cos(snapAngle)+point1.X+0.5f);
 	newPoint.Y = (int)(lineMag*sin(snapAngle)+point1.Y+0.5f);
@@ -2079,7 +2079,7 @@ ui::Point GameView::lineSnapCoords(ui::Point point1, ui::Point point2)
 ui::Point GameView::rectSnapCoords(ui::Point point1, ui::Point point2)
 {
 	ui::Point newPoint(0, 0);
-	float snapAngle = floor((atan2(point2.Y-point1.Y, point2.X-point1.X)+M_PI*0.25)/(M_PI*0.5)+0.5)*M_PI*0.5 - M_PI*0.25;
+	float snapAngle = floor((atan2((float)point2.Y-point1.Y, point2.X-point1.X)+M_PI*0.25)/(M_PI*0.5)+0.5)*M_PI*0.5 - M_PI*0.25;
 	float lineMag = sqrtf(pow((float)(point2.X-point1.X),2)+pow((float)(point2.Y-point1.Y),2));
 	newPoint.X = (int)(lineMag*cos(snapAngle)+point1.X+0.5f);
 	newPoint.Y = (int)(lineMag*sin(snapAngle)+point1.Y+0.5f);
