+++ src/simulation/elements/GEL.cpp
@@ -115,7 +115,7 @@ int Element_GEL::update(UPDATE_FUNC_ARGS)
 				dy = parts[i].y - parts[r>>8].y;
 
 				//Stickness
-				if ((dx*dx + dy*dy)>1.5 && (gel || !sim->elements[r&0xFF].Falldown || (fabs(rx)<2 && fabs(ry)<2)))
+				if ((dx*dx + dy*dy)>1.5 && (gel || !sim->elements[r&0xFF].Falldown || (fabs((float)rx)<2 && fabs((float)ry)<2)))
 				{
 					float per, nd;
 					nd = dx*dx + dy*dy - 0.5;
