+++ src/simulation/elements/SPRK.cpp
@@ -121,7 +121,7 @@ int Element_SPRK::update(UPDATE_FUNC_ARGS)
 								parts[i].life=60;
 							parts[p].temp=parts[p].life*parts[i].tmp/2.5;
 							parts[p].tmp2=1;
-							parts[p].tmp=atan2(-ry, rx)/M_PI*360;
+							parts[p].tmp=atan2(-ry, (float)rx)/M_PI*360;
 							parts[i].temp-=parts[i].tmp*2+parts[i].temp/5; // slight self-cooling
 							if (fabs(sim->pv[y/CELL][x/CELL])!=0.0f)
 							{
